memoAppクラスから、標準入力を受け取る処理を切り出しました
